+++ src/layout/Portal/index.ts
@@ -1,0 +1,3 @@
+export { PortalBanner } from './PortalBanner'
+export { PortalHeader } from './PortalHeader'
+export { PortalView } from './PortalView'
