+++ src/layout/styles/SideMenu.scss
@@ -5,7 +5,7 @@
     0px 1px 2px 0px rgba(26, 36, 126, 0.3);
   position: relative;
   min-width: $sidemenu-width;
-  z-index: 2;
+  z-index: 3;
 }
 
 .sidemenu-content {
