+++ src/screens/Project/index.ts
@@ -1,5 +1,5 @@
 export * from './AllProjects'
 export * from './ProjectDetails/ProjectDetails'
-export * from './TaskManagment'
 export * from './TeamDetails/TeamMembers'
+export * from './TaskManagement/TaskBoard/TaskBoard'
 export * from './UnassignedProject'
